comments and SkipValue changes
