+++ go/libraries/doltcore/schema/typeinfo/datetime.go
@@ -70,7 +70,7 @@ func (ti *datetimeType) ConvertNomsValueToValue(v types.Value) (interface{}, err
 	return nil, fmt.Errorf(`"%v" cannot convert NomsKind "%v" to a value`, ti.String(), v.Kind())
 }
 
-
+// ReadFrom reads a go value from a noms types.CodecReader directly
 func (ti *datetimeType) ReadFrom(_ *types.NomsBinFormat, reader types.CodecReader) (interface{}, error) {
 	k := reader.ReadKind()
 	switch k {
