+++ go/libraries/doltcore/schema/typeinfo/decimal.go
@@ -76,7 +76,7 @@ func (ti *decimalType) ConvertNomsValueToValue(v types.Value) (interface{}, erro
 	return nil, fmt.Errorf(`"%v" cannot convert NomsKind "%v" to a value`, ti.String(), v.Kind())
 }
 
-
+// ReadFrom reads a go value from a noms types.CodecReader directly
 func (ti *decimalType) ReadFrom(_ *types.NomsBinFormat, reader types.CodecReader) (interface{}, error) {
 	k := reader.ReadKind()
 	switch k {
