+++ go/libraries/doltcore/schema/typeinfo/enum.go
@@ -80,7 +80,7 @@ func (ti *enumType) ConvertNomsValueToValue(v types.Value) (interface{}, error) 
 	return nil, fmt.Errorf(`"%v" cannot convert NomsKind "%v" to a value`, ti.String(), v.Kind())
 }
 
-
+// ReadFrom reads a go value from a noms types.CodecReader directly
 func (ti *enumType) ReadFrom(_ *types.NomsBinFormat, reader types.CodecReader) (interface{}, error) {
 	k := reader.ReadKind()
 	switch k {
