+++ go/libraries/doltcore/schema/typeinfo/float.go
@@ -72,7 +72,7 @@ func (ti *floatType) ConvertNomsValueToValue(v types.Value) (interface{}, error)
 	return nil, fmt.Errorf(`"%v" cannot convert NomsKind "%v" to a value`, ti.String(), v.Kind())
 }
 
-
+// ReadFrom reads a go value from a noms types.CodecReader directly
 func (ti *floatType) ReadFrom(nbf *types.NomsBinFormat, reader types.CodecReader) (interface{}, error) {
 	k := reader.ReadKind()
 	switch k {
