+++ go/libraries/doltcore/schema/typeinfo/unknown.go
@@ -32,6 +32,8 @@ var UnknownType TypeInfo = &unknownImpl{}
 func (ti *unknownImpl) ConvertNomsValueToValue(types.Value) (interface{}, error) {
 	return nil, fmt.Errorf(`"Unknown" cannot convert any Noms value to a go value`)
 }
+
+// ReadFrom reads a go value from a noms types.CodecReader directly
 func (ti *unknownImpl) ReadFrom(_ *types.NomsBinFormat, reader types.CodecReader) (interface{}, error) {
 	return nil, fmt.Errorf(`"Unknown" cannot read any Noms value to a go value`)
 }
